Update : handling the parsing

+++ 09/ex01/RPN.hpp
@@ -11,14 +11,14 @@
 
 class RPN {
     private:
-        std::stack<double> numbers;
+        std::stack<int> numbers;
     public:
         RPN();
         RPN &operator=(const RPN &copy);
         RPN(const RPN &copy);
         ~RPN();
         //
-        double  ft_run_operator(double num2, double num1, char op);
+        int  ft_run_operator(int num2, int num1, char op);
         void    start(const std::string& input);
         bool    isOperator(char c);
 };
